fix: 반려견 프로필 카드 UI 수정
