+++ src/components/Pet/PetProfileCard.module.css
@@ -1,79 +1,47 @@
 .container {
-  position: relative;
   width: 100%;
   padding: 24px 32px;
+
   border-radius: 8px;
   background-color: var(--white);
 }
 
-.cancelButtonContainer {
-  position: absolute;
-  right: 28px;
-}
-
 .wrapper {
-  width: 100%;
   display: flex;
-  flex-direction: column;
+  flex-direction: row;
   justify-content: start;
-  align-items: start;
-  padding: 48px 32px 48px 40px;
-  gap: 36px;
-}
-
-.contentWrapper {
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  justify-content: start;
-  align-items: start;
-  gap: 20px;
+  align-items: center;
+  gap: 24px;
 }
 
 .contentRowWrapper {
-  width: 100%;
   display: flex;
   flex-direction: row;
-  gap: 20px;
+  align-items: center;
+  gap: 8px;
 }
 
-.nameText {
+.nameContent {
   font-family: 'NanumSquareNeoExtraBold';
-  font-size: 20px;
+  font-size: 16px;
   color: var(--gray-100);
-  margin-bottom: 8px;
 }
 
-.birthText {
+.breedContent {
   font-size: 14px;
   font-weight: 500;
-  color: var(--gray-60);
+  color: var(--gray-40);
 }
 
-.labelText {
-  width: 64px;
+.birthContent {
   font-size: 16px;
   font-weight: 500;
   color: var(--gray-60);
-}
-
-.contentText {
-  font-size: 16px;
-  font-weight: 500;
-  color: var(--gray-100);
-}
-
-.profileImgContainer {
-  width: 100%;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  padding-bottom: 36px;
-  border-bottom: 1px solid var(--gray-10);
 }
 
 .profileImg {
-  width: 220px;
-  border-radius: 200px;
+  width: 80px;
+  height: 80px;
+  border-radius: 80px;
   object-fit: cover;
 }
