+++ src/components/Pet/PetProfileDetailCard.module.css
@@ -1,0 +1,81 @@
+.container {
+  position: relative;
+  width: 100%;
+  padding: 24px 32px;
+  border-radius: 8px;
+  background-color: var(--white);
+}
+
+.cancelButtonContainer {
+  position: absolute;
+  right: 28px;
+}
+
+.wrapper {
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  justify-content: start;
+  align-items: start;
+  padding: 48px 32px 48px 40px;
+  gap: 36px;
+}
+
+.contentWrapper {
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  justify-content: start;
+  align-items: start;
+  gap: 20px;
+}
+
+.contentRowWrapper {
+  width: 100%;
+  display: flex;
+  flex-direction: row;
+  gap: 20px;
+}
+
+.nameText {
+  font-family: 'NanumSquareNeoExtraBold';
+  font-size: 20px;
+  color: var(--gray-100);
+  margin-bottom: 8px;
+}
+
+.birthText {
+  font-size: 14px;
+  font-weight: 500;
+  color: var(--gray-60);
+}
+
+.labelText {
+  width: 64px;
+  font-size: 16px;
+  font-weight: 500;
+  color: var(--gray-60);
+}
+
+.contentText {
+  font-size: 16px;
+  font-weight: 500;
+  color: var(--gray-100);
+}
+
+.profileImgContainer {
+  width: 100%;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  padding-bottom: 36px;
+  border-bottom: 1px solid var(--gray-10);
+}
+
+.profileImg {
+  width: 100%;
+  aspect-ratio: 1 / 1; /* 가로세로 비율 1대1 지정 */
+  max-width: 220px;
+  border-radius: 200px;
+  object-fit: cover;
+}
